# emas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  handler.SettingsResponse:
    properties:
      ai_enabled:
        type: boolean
      integrations:
        items:
          type: string
        type: array
      language:
        type: string
      notifications:
        type: boolean
      theme:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Factory operations API.
  title: eMas Factory API
  version: "1.0"
paths:
...
  /settings/get:
    get:
      consumes:
      - application/json
      description: Get settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SettingsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get settings
      tags:
      - settings
  /settings/update:
    put:
      consumes:
      - application/json
      description: Update settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SettingsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update settings
      tags:
      - settings
...
swagger: "2.0"